# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Indoor Radio System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /playlists:
    get:
      operationId: listPlaylists
      tags: [playlists]
      summary: List playlists
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaylistBody"
      responses:
        "201":
          description: Playlist created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /playlists/{playlistId}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get a playlist with its items
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistWithItems"
    put:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Update a playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaylistBody"
      responses:
        "200":
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /playlists/{playlistId}/items:
    post:
      operationId: addPlaylistItem
      tags: [playlists]
      summary: Add a media item to a playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPlaylistItemBody"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistItem"
  /playlists/{playlistId}/items/batch:
    post:
      operationId: addPlaylistItemsBatch
      tags: [playlists]
      summary: Add multiple media items to a playlist at once
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPlaylistItemsBatchBody"
      responses:
        "201":
          description: Items added
          content:
            application/json:
              schema:
                type: object
                properties:
                  added:
                    type: integer
                required: [added]
  /playlists/{playlistId}/reorder:
    put:
      operationId: reorderPlaylistItems
      tags: [playlists]
      summary: Reorder playlist items (drag and drop)
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderItemsBody"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /playlists/{playlistId}/items/{itemId}:
    delete:
      operationId: removePlaylistItem
      tags: [playlists]
      summary: Remove an item from a playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /reports/playbacks:
    get:
      operationId: getPlaybackReport
      tags: [reports]
      summary: Get playback execution report
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: clientEmail
          in: query
          schema:
            type: string
        - name: mediaId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Playback report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackReport"
  /reports/client-sessions:
    get:
      operationId: getClientSessionsReport
      tags: [reports]
      summary: Per-client report of player sessions and jingle plays per email
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sessions and jingle plays grouped per email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientSessionsReport"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    MediaItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        type:
          type: string
          enum: [music, jingle]
        filename:
          type: string
        url:
          type: string
        duration:
          type: number
          description: Duration in seconds
        clientId:
          type: integer
        gain:
          type: number
          description: Volume gain preset (0.0 to 2.0)
        coverUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - filename
        - url
        - clientId
        - createdAt
...
    Playlist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        playbackMode:
          type: string
          enum: [sequential, shuffle]
        itemCount:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - clientId
        - active
        - createdAt
    PlaylistWithItems:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: integer
        playbackMode:
          type: string
          enum: [sequential, shuffle]
        active:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - clientId
        - active
        - items
        - createdAt
    PlaylistItem:
      type: object
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        mediaId:
          type: integer
        position:
          type: integer
        media:
          $ref: "#/components/schemas/MediaItem"
      required:
        - id
        - playlistId
        - mediaId
        - position
    CreatePlaylistBody:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: integer
        playbackMode:
          type: string
          enum: [sequential, shuffle]
      required:
        - name
        - clientId
    UpdatePlaylistBody:
      type: object
      properties:
        name:
          type: string
        playbackMode:
          type: string
          enum: [sequential, shuffle]
        active:
          type: boolean
    AddPlaylistItemBody:
      type: object
      properties:
        mediaId:
          type: integer
        position:
          type: integer
      required:
        - mediaId
    AddPlaylistItemsBatchBody:
      type: object
      properties:
        mediaIds:
          type: array
          items:
            type: integer
          description: Array of media IDs to add at the end of the playlist
      required:
        - mediaIds
    ReorderItemsBody:
      type: object
      properties:
        itemIds:
          type: array
          items:
            type: integer
          description: Ordered array of playlist item IDs
      required:
        - itemIds
...
    ClientSessionsReport:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/PlayerSession"
        emailSummary:
          type: array
          items:
            $ref: "#/components/schemas/EmailSessionSummary"
      required:
        - clientId
        - clientName
        - sessions
        - emailSummary
    PlayerSession:
      type: object
      properties:
        email:
          type: string
        deviceUuid:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        durationMinutes:
          type: number
        jinglePlays:
          type: integer
        musicPlays:
          type: integer
      required:
        - email
        - deviceUuid
        - startedAt
        - endedAt
        - durationMinutes
        - jinglePlays
        - musicPlays
    EmailSessionSummary:
      type: object
      properties:
        email:
          type: string
        sessionsCount:
          type: integer
        totalDurationMinutes:
          type: number
        jingles:
          type: array
          items:
            $ref: "#/components/schemas/EmailJingleStat"
      required:
        - email
        - sessionsCount
        - totalDurationMinutes
        - jingles
    EmailJingleStat:
      type: object
      properties:
        mediaId:
          type: integer
        title:
          type: string
        plays:
          type: integer
      required:
        - mediaId
        - title
        - plays
    PlaybackReport:
      type: object
      properties:
        totalPlays:
          type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/PlaybackLogEntry"
      required:
        - totalPlays
        - entries
    PlaybackLogEntry:
      type: object
      properties:
        id:
          type: integer
        mediaId:
          type: integer
        mediaTitle:
          type: string
        mediaType:
          type: string
          enum: [music, jingle]
        deviceUuid:
          type: string
        clientEmail:
          type: string
        playedAt:
          type: string
          format: date-time
      required:
        - id
        - mediaId
        - mediaTitle
        - mediaType
        - deviceUuid
        - clientEmail
        - playedAt
...